Introduce yaml output format
